Automation of Docker, Jenkins, Git, Github, Rhel8.
This project is developed by seeing the problems faced by the Developers and the Operation Guys in the industry.
Problem- Developers create Machine Learning Model but for getting more accuracy they need to update the model again and again manually, then further checks the accuracy. This process takes so much time and effort to reduce this time and effort, this project had created.

In this project the work which developer does through manually can be done through automation of Docker and Jenkins, Jenkins automatically first pull the code from Github, checks the accuracy condition, tweaks the code and upload the code to Github. This process keep on going until the Model does not get the required accuracy.